Add files via upload
In Phase 1 of the AI Chatbot project under the Naan Mudhalvan program, the primary objective was to design and develop the basic framework and functionality of an intelligent chatbot system. The chatbot is intended to provide quick and relevant responses to user queries related to academic resources, course details, placement assistance, and other student-related services within an educational institution.